[Refactor] Make node badge a vue component
