+++ src/extensions/core/index.ts
@@ -21,4 +21,3 @@ import './uploadImage'
 import './webcamCapture'
 import './widgetInputs'
 import './uploadAudio'
-import './nodeBadge'
